Refatorando metodo de leitura do XLS das enquetes para suportar grupamentos

+++ KoboApiService.py
@@ -40,7 +40,7 @@ def retorna_key_do_formulario(u_r_l, id, u_s_e_r, p_a_s_s_w_o_r_d):
 
 def retorna_lista_com_labels(u_r_l, i_d, u_s_e_r, p_a_s_s_w_o_r_d):
     id_string = retorna_key_do_formulario('https://' + u_r_l + '/api/v1/data', i_d, u_s_e_r, p_a_s_s_w_o_r_d)
-    formulario = importar_xls_para_lista_filtrada('https://' + u_r_l + '/api/v1/data/' + str(i_d) + '.xls', u_s_e_r, p_a_s_s_w_o_r_d)
+    formulario = importar_xls_grupamento_para_lista('https://' + u_r_l + '/api/v1/data/' + str(i_d) + '.xls', u_s_e_r, p_a_s_s_w_o_r_d)
     if u_r_l == 'kobocat.docker.kobo.techo.org':
         u_r_l = 'koboform.docker.kobo.techo.org'
     if u_r_l == 'kc.humanitarianresponse.info':
@@ -58,33 +58,70 @@ def retorna_lista_com_labels(u_r_l, i_d, u_s_e_r, p_a_s_s_w_o_r_d):
     retorno = [formulario, id_string]
     return retorno
 
-def importar_xls_para_lista_filtrada(u_r_l, u_s_e_r, p_a_s_s_w_o_r_d):
+
+def importar_xls_grupamento_para_lista(u_r_l, u_s_e_r, p_a_s_s_w_o_r_d):
     req = requests.get(u_r_l, auth=(u_s_e_r, p_a_s_s_w_o_r_d))
     file = "formulario.xls"
     with open('formulario.xls', 'wb') as output:
         output.write(req.content)
     workbook = xlrd.open_workbook('formulario.xls')
 
-    worksheet = workbook.sheet_by_index(0)
-    total_linhas = worksheet.nrows
-    total_colunas = worksheet.ncols
-
+    main_sheet = workbook.sheet_by_index(0)
+    main_linhas = main_sheet.nrows
+    main_colunas = main_sheet.ncols
     table = list()
     record = dict()
 
-    for x in range(1, total_linhas):
-        for y in range(total_colunas):
-            if worksheet.cell(x, y).value:
-                if '/' in worksheet.cell(0, y).value:
-                    a = worksheet.cell(0, y).value.split('/')
-                    record[a[len(a) - 1]] = worksheet.cell(x, y).value
+    for x in range(1, main_linhas):
+        for y in range(main_colunas):
+            if main_sheet.cell(x, y).value and 'id' not in main_sheet.cell(0, y).value \
+                    and 'ObjectId' not in main_sheet.cell(0, y).value:
+                if '/' in main_sheet.cell(0, y).value:
+                    a = main_sheet.cell(0, y).value.split('/')
+                    record[a[len(a) - 1]] = main_sheet.cell(x, y).value
                 else:
-                    record[worksheet.cell(0, y).value] = worksheet.cell(x, y).value
+                    record[main_sheet.cell(0, y).value] = main_sheet.cell(x, y).value
         if record:
             table.append(record)
-    try:
-        os.remove(file)
-    except OSError as e:
-        print("Error %s - %s" % (e.filename, e.strerror))
-    return table
+    total = 1
+    if workbook.nsheets < 2:
+        try:
+            os.remove(file)
+        except OSError as e:
+            print("Error %s - %s" % (e.filename, e.strerror))
+        return table
+    else:
+        while total <= workbook.nsheets - 1:
+            linhas = workbook.sheet_by_index(total).nrows
+            colunas = workbook.sheet_by_index(total).ncols
+            table_grupamento = list()
+            for item in table:
+                item[workbook.sheet_by_index(total).name] = {}
+            record = dict()
+            for x in range(1, linhas):
+                for y in range(colunas):
+                    if workbook.sheet_by_index(total).cell(x, y).value and 'id' not in workbook.sheet_by_index(
+                            total).cell(0, y).value \
+                            and 'ObjectId' not in main_sheet.cell(0, y).value:
+                        if '/' in workbook.sheet_by_index(total).cell(0, y).value:
+                            a = workbook.sheet_by_index(total).cell(0, y).value.split('/')
+                            record[a[len(a) - 1]] = workbook.sheet_by_index(total).cell(x, y).value
+                        else:
+                            record[workbook.sheet_by_index(total).cell(0, y).value] \
+                                = workbook.sheet_by_index(total).cell(x, y).value
+                if record:
+                    table_grupamento.append(record)
+                for ele in table_grupamento:
+                    for element in table:
+                        if element['_index'] == ele['_parent_index']:
+                            for key, value in ele.items():
+                                if 'parent' not in key and 'id' not in key:
+                                    key_value = dict()
+                                    key_value[key] = value
+                                    element[workbook.sheet_by_index(total).name] = key_value
+            total = total + 1
 
+        return table
+
+
+print(retorna_respostas_com_labels('kc.humanitarianresponse.info', 134046, 'riodejaneiro', 'teto2015'))
